requirements.txt
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.1.0
websocket-client==1.6.4
wsproto==1.2.0
